no error on empty uuid
For some reason the previous code chokes on an IllegalArgumentException from the UUID parser.
Since parsing a UUID is trivial but a bit time consuming it'd be nice to store the UUIDs in their proper data type from the start.
Then, although it's not exactly beautiful, we can log out the parsing exception as soon as it occurs and null the field (or rather keep the implicit null).

After this code needs to be aware of the nullable return type, but that shouldn't be an issue since the only place in which the uuid is used it is passed directly to the equals method of another UUID.

Signed-off-by: benaryorg <binary@benary.org>