Major update to German lang file
Couldn't help but notice how out-of-date the German lang file was, so I took it upon myself to fix it. I don't speak German, so there are bound to be some errors here, but I'm confident that it's a big improvement nonetheless. I shall make sure to update this file accordingly.

• Complete update of the lang file, making it match the more up-to-date files